requirements.txt
streamlit
openai-whisper
gtts
huggingface_hub
audio-recorder-streamlit
soundfile
numpy
torch